Add chempot map
